refactor: standardize formatDate import to lib/utils across all files
